Fixed return type (array -> collection), rename Match -> Coincidence(reasone is a reserved keyword since PHP 8.0 )
