+++ backend/app/Models/Coincidence.php
@@ -6,7 +6,7 @@ namespace App\Models;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
-final class Match extends Model
+final class Coincidence extends Model
 {
     use HasFactory;
