+++ backend/app/Providers/RepositoryServiceProvider.php
@@ -10,8 +10,8 @@ use App\Repositories\EventType\EventTypeRepository;
 use App\Repositories\EventType\EventTypeRepositoryInterface;
 use App\Repositories\Location\LocationRepository;
 use App\Repositories\Location\LocationRepositoryInterface;
-use App\Repositories\Match\MatchRepository;
-use App\Repositories\Match\MatchRepositoryInterface;
+use App\Repositories\Coincidence\MatchRepository;
+use App\Repositories\Coincidence\MatchRepositoryInterface;
 use App\Repositories\Role\RoleRepository;
 use App\Repositories\Role\RoleRepositoryInterface;
 use App\Repositories\User\UserRepository;
