+++ backend/app/Repositories/Coincidence/CoincidenceRepositoryInterface.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-namespace App\Repositories\UserMedia;
+namespace App\Repositories\Coincidence;
 
 use Illuminate\Database\Eloquent\Collection;
 
-interface UserMediaRepositoryInterface
+interface CoincidenceRepositoryInterface
 {
     public function findAll(): Collection;
 }
