+++ backend/app/Repositories/Location/LocationRepository.php
@@ -6,13 +6,14 @@ namespace App\Repositories\Location;
 
 use App\Models\Location;
 use App\Repositories\BaseRepository;
+use Illuminate\Database\Eloquent\Collection;
 
 final class LocationRepository extends BaseRepository implements LocationRepositoryInterface
 {
 
-    public function findAll(): array
+    public function findAll(): Collection
     {
-        return Location::all()->toArray();
+        return Location::all();
     }
 
     public function getById($id): Location
