+++ backend/app/Repositories/Location/LocationRepositoryInterface.php
@@ -5,9 +5,10 @@ declare(strict_types=1);
 namespace App\Repositories\Location;
 
 use App\Models\Location;
+use Illuminate\Database\Eloquent\Collection;
 
 interface LocationRepositoryInterface
 {
-    public function findAll(): array;
+    public function findAll(): Collection;
     public function getById(int $id): Location;
 }
