--- backend/app/Repositories/Match/MatchRepository.php
@@ -1,17 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Repositories\Match;
-
-use App\Models\Match;
-use App\Repositories\BaseRepository;
-
-final class MatchRepository extends BaseRepository implements MatchRepositoryInterface
-{
-
-    public function findAll(): array
-    {
-        return Match::all()->toArray();
-    }
-}
