--- backend/app/Repositories/Match/MatchRepositoryInterface.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Repositories\Match;
-
-interface MatchRepositoryInterface
-{
-    public function findAll(): array;
-}
