+++ backend/app/Repositories/Role/RoleRepositoryInterface.php
@@ -4,7 +4,9 @@ declare(strict_types=1);
 
 namespace App\Repositories\Role;
 
+use Illuminate\Database\Eloquent\Collection;
+
 interface RoleRepositoryInterface
 {
-    public function all(): array;
+    public function all(): Collection;
 }
