+++ backend/app/Repositories/UserInterest/UserInterestRepositoryInterface.php
@@ -4,7 +4,9 @@ declare(strict_types=1);
 
 namespace App\Repositories\UserInterest;
 
+use Illuminate\Database\Eloquent\Collection;
+
 interface UserInterestRepositoryInterface
 {
-    public function findAll(): array;
+    public function findAll(): Collection;
 }
