+++ backend/app/Repositories/UserMedia/UserMediaRepository.php
@@ -6,12 +6,13 @@ namespace App\Repositories\UserMedia;
 
 use App\Models\UserMedia;
 use App\Repositories\BaseRepository;
+use Illuminate\Database\Eloquent\Collection;
 
 final class UserMediaRepository extends BaseRepository implements UserMediaRepositoryInterface
 {
 
-    public function findAll(): array
+    public function findAll(): Collection
     {
-        return UserMedia::all()->toArray();
+        return UserMedia::all();
     }
 }
