+++ backend/app/Repositories/UserParameter/UserParameterRepository.php
@@ -6,12 +6,13 @@ namespace App\Repositories\UserParameter;
 
 use App\Models\UserParameter;
 use App\Repositories\BaseRepository;
+use Illuminate\Database\Eloquent\Collection;
 
 final class UserParameterRepository extends BaseRepository implements UserParameterRepositoryInterface
 {
 
-    public function findAll(): array
+    public function findAll(): Collection
     {
-        return UserParameter::all()->toArray();
+        return UserParameter::all();
     }
 }
