+++ backend/app/Repositories/UserParameter/UserParameterRepositoryInterface.php
@@ -4,7 +4,9 @@ declare(strict_types=1);
 
 namespace App\Repositories\UserParameter;
 
+use Illuminate\Database\Eloquent\Collection;
+
 interface UserParameterRepositoryInterface
 {
-    public function findAll(): array;
+    public function findAll(): Collection;
 }
