+++ backend/app/Repositories/UserSetting/UserSettingRepository.php
@@ -6,12 +6,13 @@ namespace App\Repositories\UserSetting;
 
 use App\Models\UserSetting;
 use App\Repositories\BaseRepository;
+use Illuminate\Database\Eloquent\Collection;
 
 final class UserSettingRepository extends BaseRepository implements UserSettingRepositoryInterface
 {
 
-    public function findAll(): array
+    public function findAll(): Collection
     {
-        return UserSetting::all()->toArray();
+        return UserSetting::all();
     }
 }
