+++ backend/app/Repositories/UserSetting/UserSettingRepositoryInterface.php
@@ -4,7 +4,9 @@ declare(strict_types=1);
 
 namespace App\Repositories\UserSetting;
 
+use Illuminate\Database\Eloquent\Collection;
+
 interface UserSettingRepositoryInterface
 {
-    public function findAll(): array;
+    public function findAll(): Collection;
 }
